backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-dotenv==1.0.1
pydantic-settings==2.7.1
email-validator==2.2.0

# Data Scraping
playwright==1.49.1
selenium==4.27.1
beautifulsoup4==4.12.3
lxml==5.3.0

# External APIs - AI Providers (Birini seç!)
openai==1.59.7  # GPT-3.5/4, GPT-4 Vision
groq>=0.4.0  # BEDAVA ve hızlı! (Önerilen)
httpx>=0.27.0  # Groq için gerekli
anthropic>=0.43.0  # Claude (opsiyonel)
huggingface-hub>=0.20.0  # Hugging Face (opsiyonel)

# Maps & Geocoding
googlemaps>=4.10.0  # Google Maps geocodingnts

# Image Processing
pillow==11.0.0
opencv-python==4.10.0.84

# HTTP & API Clients
httpx==0.28.1
aiohttp==0.11.11
requests==2.32.3

# Data Processing
pandas==2.2.3
numpy==2.2.1

# Email
sendgrid==6.11.0
resend>=2.4.0

# Celery & Redis (Background Tasks)
celery==5.4.0
redis==5.2.1

# Utils
python-slugify==8.0.4
pycountry==24.6.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
